fix: move usesidebar hook to a seperate file to resolve issue with fast refresh during dev
