+++ host/src/hooks/use-sidebar.tsx
@@ -1,0 +1,12 @@
+import { SidebarContext } from "@/lib/utils";
+import * as React from "react";
+
+
+export function useSidebar() {
+  const context = React.useContext(SidebarContext);
+  if (!context) {
+    throw new Error("useSidebar must be used within a SidebarProvider.");
+  }
+
+  return context;
+}
